loading+error messages, show char sheet after generate
